pyproject.toml
[project]
name = "axlelore"
version = "0.1.0"
description = "Local LLM-powered vehicle assistant for Raspberry Pi"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "AxleLore Team" }
]
keywords = ["automotive", "llm", "raspberry-pi", "obd2", "vehicle-assistant"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.0",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.3.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "obd>=0.7.1",
    "pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
]

build = [
    "scrapy>=2.11.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    "pymupdf>=1.23.0",
    "pytesseract>=0.3.10",
    "langchain>=0.1.0",
    "langchain-text-splitters>=0.0.1",
]

[project.urls]
"Homepage" = "https://github.com/axlelore/axlelore"
"Bug Reports" = "https://github.com/axlelore/axlelore/issues"
"Source" = "https://github.com/axlelore/axlelore"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"
